# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the documentation for your tasks list. It shows all the routes and whatever you can do with this service. Once you create your account, login and click on Authorize button at the top and add your bearer toke. Be sure to add `Bearer` before your token otherwise requests would fail.",
        "title": "Tasks List Go Documentation API",
        "contact": {
            "name": "Saeem Mehmood",
            "url": "https://github.com/saim61",
            "email": "saimmahmood61@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/createTask": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Create a task as per your liking",
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a task",
                "parameters": [
                    {
                        "description": "Required create task parameters",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.CreateTaskRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/deleteTask/:id": {
            "delete": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Delete your task by its id",
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task by its id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Required task id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/editTask": {
            "patch": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Edit a task as per your liking. Add the task id and the other parameters",
                "tags": [
                    "Tasks"
                ],
                "summary": "Edit a task",
                "parameters": [
                    {
                        "description": "Required edit task parameters",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.Task"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/editTaskStatus": {
            "patch": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Edit a task status",
                "tags": [
                    "Tasks"
                ],
                "summary": "Edit a task status",
                "parameters": [
                    {
                        "description": "Required edit task status parameters",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.EditTaskStatusRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login by using your email and password and get your token",
                "tags": [
                    "User"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Required user parameters",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.LoginSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register yourself using your email and password",
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Required user parameters",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/task/:id": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Retreive your task by its id",
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a task by its id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Required task id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tasks.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Get and view all the tasks by all users in this route.",
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks list of all users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tasks.Task"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Fetch your details by using your email and password",
                "tags": [
                    "User"
                ],
                "summary": "Get an existing user",
                "parameters": [
                    {
                        "description": "Required user parameters",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user_tasks": {
            "get": {
                "security": [
                    {
                        "bearerToken": []
                    }
                ],
                "description": "Get and view all the tasks created by you.",
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks list of current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your CSRF token. Access the GET /protected route to get it",
                        "name": "X-CSRF-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tasks.Task"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "tasks.CreateTaskRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "tasks.EditTaskStatusRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "tasks.Task": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.UserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "errorString": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.LoginSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}